# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeGoalBody {
  /** The user's goal to analyze */
  goal: string;
  /** Whether to apply brutal honesty mode */
  brutalHonesty: boolean;
}

/**
 * The feasibility rating of the goal
 */
export type AnalysisResultFeasibility =
  (typeof AnalysisResultFeasibility)[keyof typeof AnalysisResultFeasibility];

export const AnalysisResultFeasibility = {
  Realistic: "Realistic",
  Risky: "Risky",
  Unrealistic: "Unrealistic",
} as const;

export interface AnalysisResult {
  /** The feasibility rating of the goal */
  feasibility: AnalysisResultFeasibility;
  /** Reasoning behind the feasibility rating */
  reason: string;
  /** A suggested realistic plan */
  plan: string;
}

export interface ErrorResponse {
  error: string;
}